fixed prints

+++ src/prints.c
@@ -1,5 +1,7 @@
 #include "prints.h"
 
+#include "../../c_utils/src/debug.h"
+
 void print_general(const General* general) {
 	printf("General:\n\n");
 	printf("Audio Filename: %s\n", general->audioFilename);
@@ -20,7 +22,8 @@ void print_general(const General* general) {
 	printf("Countdown Offset: %d\n", general->countdownOffset);
 	printf("Special Style: %d\n", general->specialStyle);
 	printf("Widescreen Storyboard: %d\n", general->widescreenStoryboard);
-	printf("Samples Match Playback Rate: %d\n\n", general->samplesMatchPlaybackRate);
+	printf("Samples Match Playback Rate: %d\n\n",
+		   general->samplesMatchPlaybackRate);
 }
 
 void print_editor(const Editor* editor) {
@@ -99,11 +102,11 @@ void print_events(const List* events) {
 		printf("Invalid Event list.\n");
 		return;
 	}
-	int size = events->size;
-	printf("Events (%d):\n\n", size);
-	for (int i = 0; i < size; i++) {
+	size_t size = events->size;
+	printf("Events (%zu):\n\n", size);
+	for (size_t i = 0; i < size; i++) {
 		Event* event = (Event*)events->elements[i];
-		printf("(%i)\n", i);
+		printf("(%zu)\n", i);
 		print_event(event);
 		printf("\n");
 	}
@@ -126,11 +129,11 @@ void print_timingPoints(const List* timingPoints) {
 		printf("Invalid TimingPoint list.\n");
 		return;
 	}
-	int size = timingPoints->size;
-	printf("TimingPoints (%d):\n\n", size);
-	for (int i = 0; i < size; i++) {
+	size_t size = timingPoints->size;
+	printf("TimingPoints (%zu):\n\n", size);
+	for (size_t i = 0; i < size; i++) {
 		TimingPoint* timingPoint = (TimingPoint*)timingPoints->elements[i];
-		printf("(%i)\n", i);
+		printf("(%zu)\n", i);
 		print_timingPoint(timingPoint);
 		printf("\n");
 	}
@@ -142,7 +145,9 @@ void print_beatmapColour(const BeatmapColour* beatmapColour) {
 	printf("Red: %d\n", beatmapColour->red);
 	printf("Green: %d\n", beatmapColour->green);
 	printf("Blue: %d\n", beatmapColour->blue);
-	if (beatmapColour->type == 0) printf("Combo: %d\n", ((BeatmapComboColour*)beatmapColour->object)->combo);
+	if (beatmapColour->type == 0)
+		printf("Combo: %d\n",
+			   ((BeatmapComboColour*)beatmapColour->object)->combo);
 }
 
 void print_beatmapColours(const List* beatmapColours) {
@@ -150,11 +155,12 @@ void print_beatmapColours(const List* beatmapColours) {
 		printf("Invalid BeatmapColour list.\n");
 		return;
 	}
-	int size = beatmapColours->size;
-	printf("BeatmapColours (%d):\n\n", size);
-	for (int i = 0; i < size; i++) {
-		BeatmapColour* beatmapColour = (BeatmapColour*)beatmapColours->elements[i];
-		printf("(%i)\n", i);
+	size_t size = beatmapColours->size;
+	printf("BeatmapColours (%zu):\n\n", size);
+	for (size_t i = 0; i < size; i++) {
+		BeatmapColour* beatmapColour =
+			(BeatmapColour*)beatmapColours->elements[i];
+		printf("(%zu)\n", i);
 		print_beatmapColour(beatmapColour);
 		printf("\n");
 	}
@@ -180,7 +186,7 @@ void print_slider(const Slider* slider) {
 	printf("Slider:\n");
 	printf("\tCurveType: %c\n", slider->curveType);
 	printf("\tCurvePoints: [");
-	for (int i = 0; i < slider->curvePoints->size; i++) {
+	for (size_t i = 0; i < slider->curvePoints->size; i++) {
 		CurvePoint* curvePoint = (CurvePoint*)slider->curvePoints->elements[i];
 		printf("(%d, %d)", curvePoint->x, curvePoint->y);
 		if (i < slider->curvePoints->size - 1) {
@@ -196,7 +202,7 @@ void print_slider(const Slider* slider) {
 	else
 		print_ilist(slider->edgeSounds);
 	printf("\tEdgeSets: [");
-	for (int i = 0; i < slider->edgeSets->size; i++) {
+	for (size_t i = 0; i < slider->edgeSets->size; i++) {
 		EdgeSet* edgeSet = (EdgeSet*)slider->edgeSets->elements[i];
 		printf("(%s, %s)", edgeSet->normalSet, edgeSet->additionSet);
 		if (i < slider->edgeSets->size - 1) {
@@ -224,7 +230,8 @@ void print_hitObject(const HitObject* hitObject) {
 	printf("NewCombo: %d\n", hitObject->new_combo);
 	printf("ComboSkip: %d\n", hitObject->combo_skip);
 	printf("HitSound:\n");
-	if (hitObject->hitSound != NULL) print_hitSound(hitObject->hitSound);
+	if (hitObject->hitSound != NULL)
+		print_hitSound(hitObject->hitSound);
 	else {
 		printf("wtf\n");
 		exit(1);
@@ -251,11 +258,11 @@ void print_hitObjects(const List* hitObjects) {
 		printf("Invalid HitObject list.\n");
 		return;
 	}
-	int size = hitObjects->size;
-	printf("HitObjects (%d):\n\n", size);
-	for (int i = 0; i < size; i++) {
+	size_t size = hitObjects->size;
+	printf("HitObjects (%zu):\n\n", size);
+	for (size_t i = 0; i < size; i++) {
 		HitObject* hitObject = (HitObject*)hitObjects->elements[i];
-		printf("(%i)\n", i);
+		printf("(%zu)\n", i);
 		print_hitObject(hitObject);
 		printf("\n");
 	}
